+++ src/main/java/de/samply/app/PatientTestDataGeneratorConst.java
@@ -12,6 +12,7 @@ public class PatientTestDataGeneratorConst {
     // Environment Variables
     public final static String TEST_DATA_FILENAME = "TEST_DATA_FILENAME";
     public final static String TEMPORAL_DIRECTORY = "TEMPORAL_DIRECTORY";
+    public final static String PROBABILITY_CONFIG_PATH = "PROBABILITY_CONFIG_PATH";
     public final static String TEST_DATA_FILENAME_TIME_FORMAT = "TEST_DATA_FILENAME_TIME_FORMAT";
     public final static String TEST_DATA_FILENAME_PREFIX = "TEST_DATA_FILENAME_PREFIX";
 
@@ -22,6 +23,7 @@ public class PatientTestDataGeneratorConst {
     public final static String TEMPORAL_DIRECTORY_SV = HEAD_SV + TEMPORAL_DIRECTORY + ":./temporal" + BOTTOM_SV;
     public final static String TEST_DATA_FILENAME_TIME_FORMAT_SV = HEAD_SV + TEST_DATA_FILENAME_TIME_FORMAT + ":yyyy-MM-dd_HH-mm-ss" + BOTTOM_SV;
     public final static String TEST_DATA_FILENAME_PREFIX_SV = HEAD_SV + TEST_DATA_FILENAME_PREFIX + ":obds" + BOTTOM_SV;
+    public final static String PROBABILITY_CONFIG_PATH_SV = HEAD_SV + PROBABILITY_CONFIG_PATH  + BOTTOM_SV;
 
 
 }