+++ src/main/java/de/samply/probabilities/ProbabilityType.java
@@ -1,0 +1,6 @@
+package de.samply.probabilities;
+
+public enum ProbabilityType {
+    GENDER,
+    AGE
+}