+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1 - 2024-03-07]
+## [0.0.1 - 2024-04-15]
 
 ### Added
 - gitignore
@@ -17,3 +17,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Test Data Generator component
 - Generate Rest service
 - Write test data in temporal file (serialization through XmlMapper)
+- Probability Config